@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.11.2
+// 	protoc-gen-go v1.35.1
+// 	protoc        v5.27.5
 // source: student.proto
 
 package __
@@ -35,11 +35,9 @@ type StudentModel struct {
 
 func (x *StudentModel) Reset() {
 	*x = StudentModel{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_student_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_student_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StudentModel) String() string {
@@ -50,7 +48,7 @@ func (*StudentModel) ProtoMessage() {}
 
 func (x *StudentModel) ProtoReflect() protoreflect.Message {
 	mi := &file_student_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -122,11 +120,9 @@ type StudentRequest struct {
 
 func (x *StudentRequest) Reset() {
 	*x = StudentRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_student_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_student_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StudentRequest) String() string {
@@ -137,7 +133,7 @@ func (*StudentRequest) ProtoMessage() {}
 
 func (x *StudentRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_student_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -207,11 +203,9 @@ type StudentResponse struct {
 
 func (x *StudentResponse) Reset() {
 	*x = StudentResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_student_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_student_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StudentResponse) String() string {
@@ -222,7 +216,7 @@ func (*StudentResponse) ProtoMessage() {}
 
 func (x *StudentResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_student_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -276,11 +270,9 @@ type StudentDetailResponse struct {
 
 func (x *StudentDetailResponse) Reset() {
 	*x = StudentDetailResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_student_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_student_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StudentDetailResponse) String() string {
@@ -291,7 +283,7 @@ func (*StudentDetailResponse) ProtoMessage() {}
 
 func (x *StudentDetailResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_student_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -334,11 +326,9 @@ type StudentCommonResponse struct {
 
 func (x *StudentCommonResponse) Reset() {
 	*x = StudentCommonResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_student_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_student_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StudentCommonResponse) String() string {
@@ -349,7 +339,7 @@ func (*StudentCommonResponse) ProtoMessage() {}
 
 func (x *StudentCommonResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_student_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -484,7 +474,7 @@ func file_student_proto_rawDescGZIP() []byte {
 }
 
 var file_student_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_student_proto_goTypes = []interface{}{
+var file_student_proto_goTypes = []any{
 	(*StudentModel)(nil),          // 0: studentpb.StudentModel
 	(*StudentRequest)(nil),        // 1: studentpb.StudentRequest
 	(*StudentResponse)(nil),       // 2: studentpb.StudentResponse
@@ -512,68 +502,6 @@ func init() { file_student_proto_init() }
 func file_student_proto_init() {
 	if File_student_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_student_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StudentModel); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_student_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StudentRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_student_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StudentResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_student_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StudentDetailResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_student_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StudentCommonResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
