@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.11.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.27.5
 // source: student.proto
 
 package __
@@ -15,8 +15,15 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	StudentService_StudentCreate_FullMethodName = "/studentpb.StudentService/StudentCreate"
+	StudentService_StudentDelete_FullMethodName = "/studentpb.StudentService/StudentDelete"
+	StudentService_StudentUpdate_FullMethodName = "/studentpb.StudentService/StudentUpdate"
+	StudentService_StudentShow_FullMethodName   = "/studentpb.StudentService/StudentShow"
+)
 
 // StudentServiceClient is the client API for StudentService service.
 //
@@ -37,8 +44,9 @@ func NewStudentServiceClient(cc grpc.ClientConnInterface) StudentServiceClient {
 }
 
 func (c *studentServiceClient) StudentCreate(ctx context.Context, in *StudentRequest, opts ...grpc.CallOption) (*StudentCommonResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(StudentCommonResponse)
-	err := c.cc.Invoke(ctx, "/studentpb.StudentService/StudentCreate", in, out, opts...)
+	err := c.cc.Invoke(ctx, StudentService_StudentCreate_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -46,8 +54,9 @@ func (c *studentServiceClient) StudentCreate(ctx context.Context, in *StudentReq
 }
 
 func (c *studentServiceClient) StudentDelete(ctx context.Context, in *StudentRequest, opts ...grpc.CallOption) (*StudentCommonResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(StudentCommonResponse)
-	err := c.cc.Invoke(ctx, "/studentpb.StudentService/StudentDelete", in, out, opts...)
+	err := c.cc.Invoke(ctx, StudentService_StudentDelete_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -55,8 +64,9 @@ func (c *studentServiceClient) StudentDelete(ctx context.Context, in *StudentReq
 }
 
 func (c *studentServiceClient) StudentUpdate(ctx context.Context, in *StudentRequest, opts ...grpc.CallOption) (*StudentCommonResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(StudentCommonResponse)
-	err := c.cc.Invoke(ctx, "/studentpb.StudentService/StudentUpdate", in, out, opts...)
+	err := c.cc.Invoke(ctx, StudentService_StudentUpdate_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -64,8 +74,9 @@ func (c *studentServiceClient) StudentUpdate(ctx context.Context, in *StudentReq
 }
 
 func (c *studentServiceClient) StudentShow(ctx context.Context, in *StudentRequest, opts ...grpc.CallOption) (*StudentDetailResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(StudentDetailResponse)
-	err := c.cc.Invoke(ctx, "/studentpb.StudentService/StudentShow", in, out, opts...)
+	err := c.cc.Invoke(ctx, StudentService_StudentShow_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -74,7 +85,7 @@ func (c *studentServiceClient) StudentShow(ctx context.Context, in *StudentReque
 
 // StudentServiceServer is the server API for StudentService service.
 // All implementations must embed UnimplementedStudentServiceServer
-// for forward compatibility
+// for forward compatibility.
 type StudentServiceServer interface {
 	StudentCreate(context.Context, *StudentRequest) (*StudentCommonResponse, error)
 	StudentDelete(context.Context, *StudentRequest) (*StudentCommonResponse, error)
@@ -83,9 +94,12 @@ type StudentServiceServer interface {
 	mustEmbedUnimplementedStudentServiceServer()
 }
 
-// UnimplementedStudentServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedStudentServiceServer struct {
-}
+// UnimplementedStudentServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedStudentServiceServer struct{}
 
 func (UnimplementedStudentServiceServer) StudentCreate(context.Context, *StudentRequest) (*StudentCommonResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method StudentCreate not implemented")
@@ -100,6 +114,7 @@ func (UnimplementedStudentServiceServer) StudentShow(context.Context, *StudentRe
 	return nil, status.Errorf(codes.Unimplemented, "method StudentShow not implemented")
 }
 func (UnimplementedStudentServiceServer) mustEmbedUnimplementedStudentServiceServer() {}
+func (UnimplementedStudentServiceServer) testEmbeddedByValue()                        {}
 
 // UnsafeStudentServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to StudentServiceServer will
@@ -109,6 +124,13 @@ type UnsafeStudentServiceServer interface {
 }
 
 func RegisterStudentServiceServer(s grpc.ServiceRegistrar, srv StudentServiceServer) {
+	// If the following call pancis, it indicates UnimplementedStudentServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&StudentService_ServiceDesc, srv)
 }
 
@@ -122,7 +144,7 @@ func _StudentService_StudentCreate_Handler(srv interface{}, ctx context.Context,
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/studentpb.StudentService/StudentCreate",
+		FullMethod: StudentService_StudentCreate_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(StudentServiceServer).StudentCreate(ctx, req.(*StudentRequest))
@@ -140,7 +162,7 @@ func _StudentService_StudentDelete_Handler(srv interface{}, ctx context.Context,
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/studentpb.StudentService/StudentDelete",
+		FullMethod: StudentService_StudentDelete_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(StudentServiceServer).StudentDelete(ctx, req.(*StudentRequest))
@@ -158,7 +180,7 @@ func _StudentService_StudentUpdate_Handler(srv interface{}, ctx context.Context,
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/studentpb.StudentService/StudentUpdate",
+		FullMethod: StudentService_StudentUpdate_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(StudentServiceServer).StudentUpdate(ctx, req.(*StudentRequest))
@@ -176,7 +198,7 @@ func _StudentService_StudentShow_Handler(srv interface{}, ctx context.Context, d
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/studentpb.StudentService/StudentShow",
+		FullMethod: StudentService_StudentShow_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(StudentServiceServer).StudentShow(ctx, req.(*StudentRequest))
