@@ -4,7 +4,7 @@ import (
 	"flag"
 	"fmt"
 	"gocache/config"
-	"gocache/internal/middleware/etcd/discovery/discovery3"
+	"gocache/discovery"
 	"gocache/internal/pkg/student/dao"
 	"gocache/internal/service"
 	"gocache/utils/logger"
@@ -23,7 +23,7 @@ func main() {
 	gm := service.NewGroupManager([]string{"scores", "website"}, serviceAddr)
 
 	//通过通信来共享内存而不是通过共享内存来通信
-	updateChan := make(chan bool)
+	updateChan := make(chan struct{})
 	svr, err := service.NewServer(updateChan, serviceAddr)
 	if err != nil {
 		logger.LogrusObj.Errorf("acquire grpc server instance failed, %v", err)
@@ -31,9 +31,9 @@ func main() {
 		return
 	}
 
-	go discovery3.DynamicServices(updateChan, config.Conf.Services["groupcache"].Name)
+	go discovery.DynamicServices(updateChan, config.Conf.Services["groupcache"].Name)
 
-	peers, err := discovery3.ListServicePeers(config.Conf.Services["groupcache"].Name)
+	peers, err := discovery.ListServicePeers(config.Conf.Services["groupcache"].Name)
 	if err != nil {
 		peers = []string{"serviceAddr"}
 	}
